+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.5
+- Update compatibility layer for v0 burns
+- Use connection pool for Redis interaction
+- Fetch public key and max confirmations from Darknode when booting
+- Fix incorrect `amount` field for v0 mints
+
 ## 0.4.4
 
 - Partial transaction persistence for gateway recovery